+++ packages/nmail_core/lib/views/email/widgets/person_card_header.dart
@@ -1,11 +1,11 @@
 import 'package:flutter/material.dart';
-import 'package:get/get.dart';
 
 import 'package:nmail_core/models/email_person.dart';
 import 'package:nmail_core/utils/email_person_utils.dart';
 
 import 'person_avatar.dart';
 import 'person_bridge_label.dart';
+import 'person_name.dart';
 
 class PersonCardHeader extends StatelessWidget {
   final EmailPerson person;
@@ -16,15 +16,12 @@ class PersonCardHeader extends StatelessWidget {
   Widget build(BuildContext context) {
     final identifier = emailPersonIdentifier(person);
     final bridgePubkey = person.bridgePubkey;
-    final name = person.pubkey == null
-        ? Text(emailPersonName(person))
-        : Obx(() => Text(emailPersonName(person)));
     final showIdentifier =
         person.pubkey != null || emailPersonName(person) != identifier;
 
     return ListTile(
       leading: PersonAvatar(person: person, radius: 24),
-      title: name,
+      title: PersonName(person: person),
       subtitle: showIdentifier || bridgePubkey != null
           ? Column(
               crossAxisAlignment: CrossAxisAlignment.start,